+++ packages/core/src/batch/BatchShaderGenerator.ts
@@ -1,5 +1,6 @@
 import { Shader } from '../shader/Shader';
-import { Program, defaultProgramTemplate } from '../shader/Program';
+import { Program } from '../shader/Program';
+import { defaultProgramGenerator } from '../shader/DefaultProgramGenerator';
 import { UniformGroup } from '../shader/UniformGroup';
 import { Matrix } from '@pixi/math';
 
@@ -67,7 +68,8 @@ export class BatchShaderGenerator
             fragmentSrc = fragmentSrc.replace(/%count%/gi, `${maxTextures}`);
             fragmentSrc = fragmentSrc.replace(/%forloop%/gi, this.generateSampleSrc(maxTextures));
 
-            this.programCache[maxTextures] = new Program(defaultProgramTemplate, { vertexSrc: this.vertexSrc, fragmentSrc });
+            this.programCache[maxTextures] = new Program(defaultProgramGenerator,
+                { vertexSrc: this.vertexSrc, fragmentSrc });
         }
 
         const uniforms = {
